+++ docs/changelog.md
@@ -78,6 +78,7 @@ Click on the version in the following list to see all changes since the previous
 * docs: add documentation to manage subtasks
 * feat: support Jira 9.3.0 and enterprise versions 8.20.13 and 8.13.26
 * feat: reload page after subtasks are created and dialog is closed (#30)
+* feat: check for valid license in subtask dialog (#6)
 
 [Unreleased]: https://github.com/codescape/jira-multiple-subtasks/compare/22.5.0...HEAD
 [22.5.0]: https://github.com/codescape/jira-multiple-subtasks/tree/22.5.0